+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The API is unstable, and [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 ### Changed
 - `parseThemeFileString` can now return `undefined` if the theme file is not valid.
 
+### Added
+- Window method `setDimensions({ innerWidth, innerHeight, outerWidth, outerHeight })` to set the size of the window.
+- Window options `innerWidth`, `innerHeight`, `outerWidth`, `outerHeight` to set the initial size of the window.
+
 ### Fixed
 - Improved accuracy of the titlebar styles, especially for tool windows.
 - Fixed active state of the titlebar buttons in Firefox.